requirements.txt
#anthropic==0.3.6
google-generativeai==0.3.1
pydantic==1.10.7
SQLAlchemy==1.4.31